+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/CHANGELOG.md
@@ -7,6 +7,11 @@
 - In `MetricsAdvisorAdministrationClient`, changed return types of sync and async methods `GetAnomalyAlertConfigurations` and `GetMetricAnomalyDetectionConfigurations` to pageables.
 - In `MetricsAdvisorAdministrationClient`, renamed parameter `alertConfigurationId` to `detectionConfigurationId` in sync and async `GetAnomalyAlertConfigurations` methods.
 - In `MetricEnrichedSeriesData`, made elements of `ExpectedValues`, `Periods`, `IsAnomaly`, `LowerBoundaries` and `UpperBoundaries` nullables.
+- Renamed `MetricDimension` to `DataFeedDimension`.
+- Renamed `DataAnomaly` to `DataPointAnomaly`.
+- Renamed `IncidentStatus` to `AnomalyIncidentStatus`.
+- Renamed `AlertingHook`, `EmailHook`, and `WebHook` to `NotificationHook`, `EmailNotificationHook`, and `WebNotificationHook`, respectively.
+- Renamed `TimeMode` to `AlertQueryTimeMode`.
 
 ## 1.0.0-beta.1 (2020-10-08)